+++ CHANGELOG.md
@@ -1,10 +1,20 @@
 # MMM-Fuel Changelog
-## [2.3.2]
 
+## [Unreleased]
+
+### Fixed
 
 ### Added
 
-* Provider: gasbuddy
+* Provider: `gasbuddy`
+* Config option: `fade`
+
+### Changed
+
+* Migrated provider `nsw` from API v1 to v2 in order to support new region Tasmania (Australia)
+
+### Removed
+
 ## [2.3.1]
 
 ### Fixed